refactor(console): split timeline into components/timeline + align icons
SessionDetail.tsx: 1724 → 809 lines (915-line trim). The waterfall
machinery (Span / Turn types, deriveSpans, bucketIntoTurns,
TimelineView, TurnCard, IdleDivider, DetailPanel) moved to a dedicated
components/timeline/ folder split into types.ts / derive.ts /
TimelineView.tsx. Event interface migrates to lib/events.ts so both
Conversation view and Timeline view can import a single source of
truth.

A new TimelineRow component encapsulates the sticky-bg /
label-column / chart-area / duration-column pattern, killing the
3× duplicated `${isSelected ? "bg-info-subtle/40" : "bg-bg-surface/30"}`
that used to live independently on each cell. Constants
LABEL_COL_W / DURATION_COL_W / SIDE_PANEL_W replace the magic
`224 + chartPx + 80` arithmetic that had to be kept in sync with
Tailwind utility classes by hand.

Icons re-aligned to the sidebar — same SVG paths as Layout.tsx so the
agent / env / vault icon in the session header is byte-identical to
the icon in the nav for the same kind. (Was: lucide-style robot /
cloud / key icons that didn't match anything else in the app.)

Also: add the missing /v1/integrations/slack/agents/:id/publications
route (parallel to the linear one already on the agent-keyed reverse
lookup path) so AgentDetail's "Published to Slack" chip stops 404'ing.
