+++ apps/console/src/pages/SessionDetail.tsx
@@ -1,37 +1,12 @@
-import { useEffect, useMemo, useState, useRef, Fragment } from "react";
+import { useEffect, useState, useRef } from "react";
 import { useParams, Link } from "react-router";
 import { useApi } from "../lib/api";
 import { Markdown } from "../components/Markdown";
-import { formatDuration, formatRelative, pickTickStep, shortenId } from "../lib/format";
+import { formatDuration, formatRelative, shortenId } from "../lib/format";
 import { Badge, StatusPill } from "../components/Badge";
 import { AgentIcon, ClockIcon, DurationIcon, EnvIcon, VaultIcon } from "../components/icons";
-
-interface Event {
-  type: string;
-  content?: Array<{ type: string; text: string }> | string;
-  id?: string;
-  name?: string;
-  input?: Record<string, unknown>;
-  tool_use_id?: string;
-  mcp_tool_use_id?: string;
-  mcp_server_name?: string;
-  error?: string;
-  source?: string;
-  message?: string;
-  stop_reason?: { type: string };
-  /** Canonical id for streamed assistant messages — set on
-   *  agent.message_stream_start / _chunk / _stream_end and on the
-   *  matching final agent.message. Lets the renderer correlate
-   *  in-flight chunks with the eventually-committed message. */
-  message_id?: string;
-  delta?: string;
-  /** ISO timestamp. Server sets it for stored events; the client tags streamed
-   *  events on arrival with Date.now() as a best-effort fallback. */
-  ts?: string;
-  /** Server-side monotonic seq. Only set for events fetched from /events. */
-  seq?: number;
-  [key: string]: unknown;
-}
+import { TimelineView } from "../components/timeline/TimelineView";
+import type { Event } from "../lib/events";
 
 type View = "chat" | "timeline";
 
@@ -831,894 +806,4 @@ function EventBubble({ event }: { event: Event }) {
   }
 }
 
-// ─── Timeline (waterfall) ────────────────────────────────────────────────
-//
-// Pure-frontend projection of the event stream into a Gantt-style timeline.
-// Tool/MCP/custom-tool calls become bars (use→result paired by id); model
-// turns are derived as the gap between the last completed event and the next
-// agent.message; messages and session.* events render as instants. We
-// deliberately drop agent.thinking (already filtered upstream) and tool
-// result events (consumed in pairing) to keep one row per logical span.
-
-type SpanFamily =
-  | "model"
-  | "tool"
-  | "mcp"
-  | "custom_tool"
-  | "user"
-  | "agent"
-  | "system"
-  | "warn"
-  | "error"
-  | "schedule"     // schedule tool's "waiting for alarm" window (10s, 1h, …)
-  | "wakeup"       // user.message synthesized by onScheduledWakeup
-  | "compaction"   // span.compaction_summarize_*
-  | "outcome"      // span.outcome_evaluation_*
-  | "thread"       // sub-agent thread lifecycle / messages
-  | "aux"          // aux.model_call (web_fetch summarizer etc.)
-  | "thinking"     // agent.thinking marker
-  | "marker";      // catch-all for unrecognized event types
-
-interface Span {
-  key: string;
-  family: SpanFamily;
-  label: string;
-  detail?: string;
-  /** ms since the first event */
-  startMs: number;
-  /** 0 for instants */
-  durationMs: number;
-  /** Optional: ms from this span's start to first-token (model spans only).
-   *  Used to render a TTFT divider inside the bar. */
-  ttftMs?: number;
-  /** Source events that contributed to this span (1 for instants,
-   *  2 for paired spans, possibly more for tool calls with streaming
-   *  input chunks). Click-to-expand renders the raw JSON of these. */
-  events: Event[];
-}
-
-const FAMILY_DOT: Record<SpanFamily, string> = {
-  model: "bg-info",
-  tool: "bg-emerald-500",
-  mcp: "bg-purple-500",
-  custom_tool: "bg-amber-500",
-  user: "bg-brand",
-  agent: "bg-fg-muted",
-  system: "bg-fg-subtle",
-  warn: "bg-warning",
-  error: "bg-danger",
-  schedule: "bg-info",
-  wakeup: "bg-info",
-  compaction: "bg-purple-400",
-  outcome: "bg-emerald-400",
-  thread: "bg-fg-muted",
-  aux: "bg-fg-subtle",
-  thinking: "bg-fg-subtle",
-  marker: "bg-fg-subtle",
-};
-
-const FAMILY_BAR: Record<SpanFamily, string> = {
-  model: "bg-info/70",
-  tool: "bg-emerald-500/70",
-  mcp: "bg-purple-500/70",
-  custom_tool: "bg-amber-500/70",
-  user: "bg-brand/70",
-  agent: "bg-fg-muted/70",
-  system: "bg-fg-subtle/70",
-  warn: "bg-warning/70",
-  error: "bg-danger/70",
-  schedule: "bg-info/40",
-  wakeup: "bg-info/70",
-  compaction: "bg-purple-400/70",
-  outcome: "bg-emerald-400/70",
-  thread: "bg-fg-muted/50",
-  aux: "bg-fg-subtle/70",
-  thinking: "bg-fg-subtle/40",
-  marker: "bg-fg-subtle/40",
-};
-
-
-function deriveSpans(events: Event[]): { spans: Span[]; totalMs: number } {
-  // Each event carries a millisecond-precision `processed_at` ISO string AND
-  // a `ts` Unix-seconds integer. The seconds-precision ts collapses every
-  // event in the same second to identical timestamps and made schedule's
-  // 10-second wait window render as a 22ms timeline. processed_at (set at
-  // SessionDO write time) is what the rest of the UI uses for ordering and
-  // is the right basis for waterfall timing too.
-  const tsMs = (e: Event): number | null => {
-    const pa = (e.data as { processed_at?: string } | undefined)?.processed_at
-      ?? (e as { processed_at?: string }).processed_at;
-    if (typeof pa === "string") {
-      const t = Date.parse(pa);
-      if (Number.isFinite(t)) return t;
-    }
-    if (typeof e.ts === "number") return e.ts * 1000; // ts is unix SECONDS — multiply
-    return null;
-  };
-
-  const timed = events.map((e) => ({ e, t: tsMs(e) })).filter((x): x is { e: Event; t: number } => x.t !== null);
-  if (timed.length === 0) return { spans: [], totalMs: 0 };
-
-  const t0 = timed[0].t;
-  const tEnd = timed[timed.length - 1].t;
-  const totalMs = Math.max(1, tEnd - t0);
-
-  const spans: Span[] = [];
-
-  // Index look-ahead pairings. O(1) instead of nested scans.
-  // Each map stores both timestamp (for span math) and the source Event
-  // (for click-to-expand JSON inspection).
-  const toolResults = new Map<string, { t: number; e: Event }>();
-  const mcpResults = new Map<string, { t: number; e: Event }>();
-  const customResults = new Map<string, { t: number; e: Event }>();
-  // model_request_start_id → end's timestamp + usage. Anthropic's wire
-  // format pairs the per-call span pair this way (rather than positional
-  // FIFO), so multiple parallel or nested model calls stay correctly
-  // associated. FIFO fallback below for events that predate the field.
-  const modelEndsById = new Map<string, { t: number; e: Event; usage?: { input_tokens: number; output_tokens: number; cache_read_input_tokens?: number; cache_creation_input_tokens?: number }; finishReason?: string }>();
-  const modelEndsFifo: { t: number; e: Event }[] = [];
-  // OMA-extension span — pairs to model_request_start the same way the end
-  // does. Lets the model bar split into TTFT (start→first_token) and
-  // generation (first_token→end). FIFO fallback for events without ids.
-  const modelFirstTokensById = new Map<string, { t: number; e: Event }>();
-  const modelFirstTokensFifo: { t: number; e: Event }[] = [];
-  const compactEnds: { t: number; e: Event }[] = [];
-  const outcomeEnds: { t: number; e: Event }[] = [];
-  // parent_event_id → child {t, event}. Used to pair span.wakeup_scheduled
-  // (parent) with its eventual user.message (child) — same EventBase field
-  // tool_result→tool_use uses, so this generalizes to any future
-  // schedule→fire / outcome→eval / etc. causal pair without needing custom
-  // id fields per kind.
-  const childByParent = new Map<string, { t: number; e: Event }>();
-  for (const { e, t } of timed) {
-    if (e.type === "agent.tool_result" && e.tool_use_id) toolResults.set(e.tool_use_id, { t, e });
-    else if (e.type === "agent.mcp_tool_result" && e.mcp_tool_use_id) mcpResults.set(e.mcp_tool_use_id, { t, e });
-    else if (e.type === "user.custom_tool_result" && (e as Event).id) customResults.set(String(e.id), { t, e });
-    else if (e.type === "span.model_request_end") {
-      modelEndsFifo.push({ t, e });
-      const data = (e.data as { model_request_start_id?: string; model_usage?: any; finish_reason?: string } | undefined);
-      const sid = (e as { model_request_start_id?: string }).model_request_start_id ?? data?.model_request_start_id;
-      if (sid) modelEndsById.set(sid, { t, e, usage: data?.model_usage, finishReason: data?.finish_reason });
-    }
-    else if (e.type === "span.model_first_token") {
-      modelFirstTokensFifo.push({ t, e });
-      const data = (e.data as { model_request_start_id?: string } | undefined);
-      const sid = (e as { model_request_start_id?: string }).model_request_start_id ?? data?.model_request_start_id;
-      if (sid) modelFirstTokensById.set(sid, { t, e });
-    }
-    else if (e.type === "span.compaction_summarize_end") compactEnds.push({ t, e });
-    else if (e.type === "span.outcome_evaluation_end") outcomeEnds.push({ t, e });
-    const pid = (e as { parent_event_id?: string }).parent_event_id
-      ?? (e.data as { parent_event_id?: string } | undefined)?.parent_event_id;
-    if (pid) childByParent.set(pid, { t, e });
-  }
-
-  // FIFO fallback indices for events that lack id-based pairing.
-  let modelEndFifoIdx = 0;
-  let modelFirstTokenFifoIdx = 0;
-  let compactEndIdx = 0;
-  let outcomeEndIdx = 0;
-
-  // Streaming chunks (deltas + start/end markers from incremental rendering)
-  // are broadcast-only; the canonical event (agent.message / agent.thinking)
-  // lands on commit and is what timeline should show.
-  const STREAMING_NOISE = new Set([
-    "agent.message_chunk",
-    "agent.message_stream_start",
-    "agent.message_stream_end",
-    "agent.thinking_chunk",
-    "agent.thinking_stream_start",
-    "agent.thinking_stream_end",
-    "agent.tool_use_input_chunk",
-    "agent.tool_use_input_stream_start",
-    "agent.tool_use_input_stream_end",
-  ]);
-
-  // Index look-back: event id → original Event so we can attach matched
-  // partner events (the `_end` of a paired span, the tool_result of a
-  // tool_use) to the resulting span row. Click-to-expand renders these.
-  const eventById = new Map<string, Event>();
-  for (const { e } of timed) {
-    const eid = (e as { id?: string }).id ?? (e.data as { id?: string } | undefined)?.id;
-    if (eid) eventById.set(eid, e);
-  }
-
-  for (let i = 0; i < timed.length; i++) {
-    const { e, t } = timed[i];
-    const startMs = t - t0;
-
-    if (STREAMING_NOISE.has(e.type)) continue;
-
-    // Default source-events list. Each push site below may extend it with
-    // the matched partner (end / result) when one exists.
-    const sourceEvents: Event[] = [e];
-    const pushSpan = (span: Omit<Span, "events">) => spans.push({ ...span, events: sourceEvents });
-
-    if (e.type === "agent.tool_use" || e.type === "agent.custom_tool_use") {
-      const result = e.type === "agent.tool_use"
-        ? toolResults.get(String(e.id))
-        : customResults.get(String(e.id));
-      const endMs = result ? result.t - t0 : startMs;
-      if (result) sourceEvents.push(result.e);
-      pushSpan({
-        key: `tool-${e.id ?? i}`,
-        family: e.type === "agent.tool_use" ? "tool" : "custom_tool",
-        label: String(e.name ?? "tool"),
-        detail: result ? "completed" : "no result",
-        startMs,
-        durationMs: Math.max(0, endMs - startMs),
-      });
-    } else if (e.type === "agent.mcp_tool_use") {
-      const result = mcpResults.get(String(e.id));
-      const endMs = result ? result.t - t0 : startMs;
-      if (result) sourceEvents.push(result.e);
-      pushSpan({
-        key: `mcp-${e.id ?? i}`,
-        family: "mcp",
-        label: `${String(e.mcp_server_name ?? "mcp")}:${String(e.name ?? "?")}`,
-        detail: result ? "completed" : "no result",
-        startMs,
-        durationMs: Math.max(0, endMs - startMs),
-      });
-    } else if (
-      e.type === "agent.tool_result" ||
-      e.type === "agent.mcp_tool_result" ||
-      e.type === "user.custom_tool_result"
-    ) {
-      // consumed via pairing above — no row
-      continue;
-    } else if (e.type === "span.model_request_start") {
-      // One pair per ai-sdk step (= one model API call). Pair via the
-      // start event id; old data without ids falls back to FIFO order.
-      const sid = String((e as { id?: string }).id ?? (e.data as { id?: string } | undefined)?.id ?? "");
-      const matched = sid ? modelEndsById.get(sid) : (modelEndsFifo[modelEndFifoIdx++] ?? undefined);
-      const ftMatch = sid ? modelFirstTokensById.get(sid) : (modelFirstTokensFifo[modelFirstTokenFifoIdx++] ?? undefined);
-      const end = matched?.t ?? t;
-      const usage = matched?.usage;
-      if (matched?.e) sourceEvents.push(matched.e);
-      if (ftMatch?.e) sourceEvents.push(ftMatch.e);
-      // Bar label gets a token count when we have it — useful at-a-glance
-      // ("did this turn read 100k tokens or 1k?") without opening details.
-      const tokSummary = usage
-        ? `${usage.input_tokens}↓ ${usage.output_tokens}↑${usage.cache_read_input_tokens ? ` ⚡${usage.cache_read_input_tokens}` : ""}`
-        : undefined;
-      const ttftMs = ftMatch ? Math.max(0, ftMatch.t - t) : undefined;
-      const ttftSummary = typeof ttftMs === "number" ? `TTFT ${formatDuration(ttftMs)}` : undefined;
-      pushSpan({
-        key: `model-${sid || i}`,
-        family: "model",
-        label: "model call",
-        detail: [matched?.finishReason, ttftSummary, tokSummary].filter(Boolean).join(" · ") || undefined,
-        startMs,
-        durationMs: Math.max(0, end - t0 - startMs),
-        ttftMs,
-      });
-    } else if (e.type === "span.compaction_summarize_start") {
-      const matched = compactEnds[compactEndIdx++];
-      const end = matched?.t ?? t;
-      if (matched?.e) sourceEvents.push(matched.e);
-      pushSpan({
-        key: `compact-${i}`,
-        family: "compaction",
-        label: "compaction",
-        startMs,
-        durationMs: Math.max(0, end - t0 - startMs),
-      });
-    } else if (e.type === "span.outcome_evaluation_start") {
-      const matched = outcomeEnds[outcomeEndIdx++];
-      const end = matched?.t ?? t;
-      if (matched?.e) sourceEvents.push(matched.e);
-      pushSpan({
-        key: `outcome-${i}`,
-        family: "outcome",
-        label: "outcome eval",
-        startMs,
-        durationMs: Math.max(0, end - t0 - startMs),
-      });
-    } else if (
-      e.type === "span.model_request_end" ||
-      e.type === "span.model_first_token" ||
-      e.type === "span.compaction_summarize_end" ||
-      e.type === "span.outcome_evaluation_end" ||
-      e.type === "span.outcome_evaluation_ongoing"
-    ) {
-      continue; // paired or progress noise
-    } else if (e.type === "span.wakeup_scheduled") {
-      // Pair via parent_event_id: the eventual wakeup user.message sets its
-      // parent_event_id to this span's id (mint-then-emit, see
-      // session-do.ts:scheduleWakeup). Bar runs scheduled → fired and
-      // visualizes the actual wait, which dwarfs everything else (10s, 1h,
-      // 1d…); without it the operator can't see the wait at all on the
-      // waterfall.
-      const sid = String((e as { id?: string }).id ?? (e.data as { id?: string } | undefined)?.id ?? "");
-      const fired = sid ? childByParent.get(sid) : undefined;
-      const endMs = fired ? fired.t - t0 : startMs;
-      if (fired?.e) sourceEvents.push(fired.e);
-      pushSpan({
-        key: `sched-${sid || i}`,
-        family: "schedule",
-        label: "schedule waiting",
-        detail: fired ? "fired" : "pending",
-        startMs,
-        durationMs: Math.max(0, endMs - startMs),
-      });
-    } else if (e.type === "user.message") {
-      const md = (e as { metadata?: { harness?: string; kind?: string } }).metadata;
-      const isWakeup = md?.harness === "schedule" && md?.kind === "wakeup";
-      pushSpan({
-        key: `u-${i}`,
-        family: isWakeup ? "wakeup" : "user",
-        label: isWakeup ? "user.message (wakeup)" : "user.message",
-        startMs,
-        durationMs: 0,
-      });
-    } else if (e.type === "agent.message") {
-      pushSpan({ key: `a-${i}`, family: "agent", label: "agent.message", startMs, durationMs: 0 });
-    } else if (e.type === "agent.thinking") {
-      pushSpan({ key: `think-${i}`, family: "thinking", label: "agent.thinking", startMs, durationMs: 0 });
-    } else if (e.type === "aux.model_call") {
-      pushSpan({ key: `aux-${i}`, family: "aux", label: "aux.model_call", startMs, durationMs: 0 });
-    } else if (
-      e.type === "agent.thread_message_sent" ||
-      e.type === "agent.thread_message_received" ||
-      e.type === "agent.thread_message" ||
-      e.type === "session.thread_created" ||
-      e.type === "session.thread_idle"
-    ) {
-      pushSpan({ key: `thread-${i}`, family: "thread", label: e.type.replace(/^.*\./, ""), startMs, durationMs: 0 });
-    } else if (e.type === "agent.thread_context_compacted") {
-      pushSpan({ key: `compact-marker-${i}`, family: "compaction", label: "thread compacted", startMs, durationMs: 0 });
-    } else if (e.type === "session.error") {
-      pushSpan({
-        key: `err-${i}`,
-        family: "error",
-        label: "session.error",
-        detail: typeof e.error === "string" ? e.error : JSON.stringify(e.error),
-        startMs,
-        durationMs: 0,
-      });
-    } else if (e.type === "session.warning") {
-      pushSpan({
-        key: `warn-${i}`,
-        family: "warn",
-        label: `warning:${String(e.source ?? "")}`,
-        detail: String(e.message ?? ""),
-        startMs,
-        durationMs: 0,
-      });
-    } else if (e.type.startsWith("session.")) {
-      pushSpan({ key: `s-${i}`, family: "system", label: e.type, startMs, durationMs: 0 });
-    } else {
-      // Catch-all: surface unknown types as instant markers rather than
-      // silently dropping. New event types added later show up immediately
-      // and the operator can decide whether to give them dedicated visuals.
-      pushSpan({ key: `mk-${i}`, family: "marker", label: e.type, startMs, durationMs: 0 });
-    }
-  }
-
-  return { spans, totalMs };
-}
-
-/**
- * A "turn" is the unit of agent work between a user trigger event
- * (user.message / user.tool_confirmation / user.custom_tool_result) and
- * the next session.status_idle / .status_terminated / .error. This is the
- * same definition the harness uses internally (see drainEventQueue in
- * apps/agent/src/runtime/session-do.ts) and matches Conversation view's
- * notion of a turn — so a Timeline burst card lines up 1:1 with a chat
- * exchange. Time-based bucketing breaks under both fast follow-ups and
- * 40-minute cron gaps; semantic bucketing handles both uniformly.
- */
-type TurnTriggerKind =
-  | "user_message"
-  | "wakeup"
-  | "tool_confirmation"
-  | "custom_tool_result"
-  | "init";
-
-type TurnStatus = "completed" | "running" | "errored" | "terminated";
-
-interface Turn {
-  id: string;
-  triggerKind: TurnTriggerKind;
-  trigger?: Event;
-  /** Wall-clock ms epoch of the trigger (or first event for init turn). */
-  triggerTs: number;
-  /** Inclusive: trigger event + everything until the closing status event. */
-  events: Event[];
-  status: TurnStatus;
-  /** Wall-clock ms epoch when the turn closed; undefined while still running. */
-  endedAt?: number;
-}
-
-function parseEventTs(e: Event): number {
-  const ts = (e as { processed_at?: string }).processed_at;
-  if (typeof ts === "string") {
-    const t = new Date(ts).getTime();
-    if (Number.isFinite(t)) return t;
-  }
-  return 0;
-}
-
-function bucketIntoTurns(events: Event[]): Turn[] {
-  const turns: Turn[] = [];
-  let current: Turn | null = null;
-
-  // user.message metadata.harness === "schedule" + kind === "wakeup" is
-  // the wire convention for cron-fired turns (see SessionDO.onScheduledWakeup).
-  // Distinguish them in the trigger badge so an operator scanning the
-  // timeline can tell at a glance "the agent woke itself" from "the user
-  // sent something".
-  const triggerKindOf = (e: Event): TurnTriggerKind | null => {
-    if (e.type === "user.message") {
-      const md = (e as { metadata?: { harness?: string; kind?: string } }).metadata;
-      if (md?.harness === "schedule" && md?.kind === "wakeup") return "wakeup";
-      return "user_message";
-    }
-    if (e.type === "user.tool_confirmation") return "tool_confirmation";
-    if (e.type === "user.custom_tool_result") return "custom_tool_result";
-    return null;
-  };
-
-  for (const e of events) {
-    const k = triggerKindOf(e);
-    if (k) {
-      current = {
-        id: `turn-${turns.length}`,
-        triggerKind: k,
-        trigger: e,
-        triggerTs: parseEventTs(e),
-        events: [e],
-        status: "running",
-      };
-      turns.push(current);
-      continue;
-    }
-    if (!current) {
-      // Pre-trigger init events (init_events injected by /init handler,
-      // platform reminders, etc.) — bucket into a synthetic init turn so
-      // they get a card rather than being silently dropped.
-      current = {
-        id: `turn-init`,
-        triggerKind: "init",
-        triggerTs: parseEventTs(e),
-        events: [e],
-        status: "running",
-      };
-      turns.push(current);
-      continue;
-    }
-    current.events.push(e);
-    if (e.type === "session.status_idle") {
-      current.status = "completed";
-      current.endedAt = parseEventTs(e);
-    } else if (e.type === "session.status_terminated") {
-      current.status = "terminated";
-      current.endedAt = parseEventTs(e);
-    } else if (e.type === "session.error") {
-      current.status = "errored";
-      current.endedAt = parseEventTs(e);
-    }
-  }
-
-  return turns;
-}
-
-const TRIGGER_LABEL: Record<TurnTriggerKind, string> = {
-  user_message: "user message",
-  wakeup: "scheduled wakeup",
-  tool_confirmation: "tool confirmation",
-  custom_tool_result: "custom tool result",
-  init: "session init",
-};
-
-const TRIGGER_DOT: Record<TurnTriggerKind, string> = {
-  user_message: "bg-brand",
-  wakeup: "bg-info",
-  tool_confirmation: "bg-amber-500",
-  custom_tool_result: "bg-amber-500",
-  init: "bg-fg-subtle",
-};
-
-const STATUS_TEXT: Record<TurnStatus, string> = {
-  completed: "text-fg-subtle",
-  running: "text-info",
-  errored: "text-danger",
-  terminated: "text-danger",
-};
-
-/** What's currently selected for the side detail panel. Lifted up to
- *  TimelineView so a click in any TurnCard updates one shared panel. */
-interface TimelineSelection {
-  spanKey: string;
-  spanLabel: string;
-  events: Event[];
-}
-
-function TimelineView({ events }: { events: Event[] }) {
-  const turns = useMemo(() => bucketIntoTurns(events), [events]);
-  const containerRef = useRef<HTMLDivElement | null>(null);
-  const [selection, setSelection] = useState<TimelineSelection | null>(null);
-
-  // Auto-scroll to the latest turn when new ones land. Skip if user has
-  // scrolled up (we can detect this with scrollHeight - scrollTop ≈ clientHeight)
-  // — they're inspecting an older turn and shouldn't get yanked.
-  useEffect(() => {
-    if (!containerRef.current) return;
-    const el = containerRef.current;
-    const nearBottom = el.scrollHeight - el.scrollTop - el.clientHeight < 200;
-    if (nearBottom) {
-      el.scrollTo({ top: el.scrollHeight, behavior: "auto" });
-    }
-  }, [turns.length]);
-
-  if (turns.length === 0) {
-    return (
-      <div className="flex-1 flex items-center justify-center text-sm text-fg-subtle">
-        No timing data yet — send a message to populate the timeline.
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex-1 flex min-h-0">
-      <div ref={containerRef} className="flex-1 min-w-0 overflow-y-auto px-8 py-6 space-y-3">
-        {turns.map((turn, i) => {
-          const prev = i > 0 ? turns[i - 1] : null;
-          const idleMs =
-            prev && prev.endedAt && turn.triggerTs && turn.triggerTs > prev.endedAt
-              ? turn.triggerTs - prev.endedAt
-              : 0;
-          return (
-            <Fragment key={turn.id}>
-              {idleMs > 0 && <IdleDivider ms={idleMs} nextKind={turn.triggerKind} />}
-              <TurnCard
-                turn={turn}
-                selection={selection}
-                onSelectSpan={(span) =>
-                  setSelection((prev) =>
-                    prev?.spanKey === span.key
-                      ? null
-                      : { spanKey: span.key, spanLabel: span.label, events: span.events },
-                  )
-                }
-              />
-            </Fragment>
-          );
-        })}
-      </div>
-      {selection && (
-        <DetailPanel
-          selection={selection}
-          onClose={() => setSelection(null)}
-        />
-      )}
-    </div>
-  );
-}
-
-function DetailPanel({
-  selection,
-  onClose,
-}: {
-  selection: TimelineSelection;
-  onClose: () => void;
-}) {
-  return (
-    <aside className="w-[420px] shrink-0 border-l border-border bg-bg flex flex-col min-h-0">
-      <div className="px-4 py-3 border-b border-border flex items-center gap-3 shrink-0">
-        <div className="min-w-0 flex-1">
-          <div className="text-[10px] uppercase tracking-wide text-fg-subtle font-mono">
-            {selection.events.length === 1
-              ? "source event"
-              : `source events (${selection.events.length})`}
-          </div>
-          <div className="text-sm font-mono text-fg-muted truncate">{selection.spanLabel}</div>
-        </div>
-        <button
-          onClick={onClose}
-          className="text-fg-subtle hover:text-fg-muted text-lg leading-none px-2"
-          title="Close"
-        >
-          ×
-        </button>
-      </div>
-      <div className="flex-1 overflow-y-auto p-4 space-y-3">
-        {selection.events.map((ev, idx) => (
-          <div key={idx} className="border border-border/60 rounded">
-            <div className="px-3 py-1.5 border-b border-border/60 bg-bg-surface/40 flex items-center gap-2 text-[11px] font-mono">
-              <span className="text-fg-muted">{ev.type}</span>
-              {typeof ev.processed_at === "string" && (
-                <span className="text-fg-subtle ml-auto">
-                  {new Date(ev.processed_at).toISOString().slice(11, 23)}
-                </span>
-              )}
-            </div>
-            <pre className="text-[11px] font-mono text-fg-muted px-3 py-2 overflow-x-auto whitespace-pre-wrap break-all">
-              {JSON.stringify(ev, null, 2)}
-            </pre>
-          </div>
-        ))}
-      </div>
-    </aside>
-  );
-}
-
-function IdleDivider({ ms, nextKind }: { ms: number; nextKind: TurnTriggerKind }) {
-  return (
-    <div className="flex items-center gap-3 text-xs text-fg-subtle font-mono py-1">
-      <div className="flex-1 border-t border-dashed border-border" />
-      <span>
-        ↓ {formatDuration(ms)} idle
-        {nextKind === "wakeup" && " · scheduled wakeup"}
-      </span>
-      <div className="flex-1 border-t border-dashed border-border" />
-    </div>
-  );
-}
-
-function TurnCard({
-  turn,
-  selection,
-  onSelectSpan,
-}: {
-  turn: Turn;
-  selection: TimelineSelection | null;
-  onSelectSpan: (span: Span) => void;
-}) {
-  const [collapsed, setCollapsed] = useState(false);
-  const { spans, totalMs } = useMemo(() => deriveSpans(turn.events), [turn.events]);
-
-  // Per-card scroll + zoom state (same shape as the old global TimelineView,
-  // now scoped to one turn — each card auto-densifies independently so a
-  // 30-min turn doesn't impose its scale on a 2-second turn next to it).
-  const scrollRef = useRef<HTMLDivElement | null>(null);
-  const [pxPerMs, setPxPerMs] = useState<number | null>(null);
-  const [mode, setMode] = useState<"auto" | "manual">("auto");
-
-  useEffect(() => {
-    if (collapsed || mode === "manual" || !scrollRef.current || totalMs <= 0) return;
-    const viewportChartPx = scrollRef.current.clientWidth - 224 - 80 - 64;
-    if (viewportChartPx <= 0) return;
-    const times: number[] = [];
-    for (const s of spans) {
-      times.push(s.startMs);
-      if (s.durationMs > 0) times.push(s.startMs + s.durationMs);
-    }
-    times.sort((a, b) => a - b);
-    const gaps: number[] = [];
-    for (let i = 1; i < times.length; i++) {
-      const g = times[i] - times[i - 1];
-      if (g > 0) gaps.push(g);
-    }
-    let candidate: number;
-    if (gaps.length === 0) {
-      candidate = viewportChartPx / totalMs;
-    } else {
-      const sorted = [...gaps].sort((a, b) => a - b);
-      const median = sorted[Math.floor(sorted.length / 2)] || 1;
-      candidate = 25 / median;
-    }
-    const auto = Math.min(5, Math.max(candidate, viewportChartPx / totalMs));
-    setPxPerMs(Math.max(auto, viewportChartPx / totalMs));
-  }, [collapsed, mode, spans, totalMs]);
-
-  const effectivePxPerMs = pxPerMs ?? 0.05;
-  const chartPx = Math.max(200, totalMs * effectivePxPerMs);
-
-  const zoomBy = (factor: number) => {
-    setMode("manual");
-    setPxPerMs((p) => Math.min(50, Math.max(0.0001, (p ?? 0.05) * factor)));
-  };
-  const fitToViewport = () => {
-    if (!scrollRef.current) return;
-    const viewportChartPx = scrollRef.current.clientWidth - 224 - 80 - 64;
-    if (viewportChartPx > 0 && totalMs > 0) setPxPerMs(viewportChartPx / totalMs);
-    setMode("manual");
-  };
-  const resetAuto = () => {
-    setMode("auto");
-    setPxPerMs(null);
-  };
-
-  // Tick spacing: aim for ~120px between labels at the current pixel
-  // density. pickTickStep takes a "total span across 6 ticks" arg, so
-  // multiply the target step by 6 to get a step matching ~120px gaps.
-  const targetTickPx = 120;
-  const desiredStepMs = targetTickPx / effectivePxPerMs;
-  const tickStep = pickTickStep(desiredStepMs * 6);
-  const ticks: number[] = [];
-  for (let t = 0; t <= totalMs; t += tickStep) ticks.push(t);
-
-  const fmtRate = (ppms: number) => {
-    const pps = ppms * 1000;
-    if (pps >= 100) return `${Math.round(pps)} px/s`;
-    if (pps >= 1) return `${pps.toFixed(1)} px/s`;
-    return `${pps.toFixed(2)} px/s`;
-  };
-
-  // Aggregate per-turn cost / token totals from span events that carry
-  // model_usage. Cheap walk — span events are already in memory.
-  const tokens = useMemo(() => {
-    let input = 0;
-    let output = 0;
-    let cacheRead = 0;
-    let calls = 0;
-    for (const e of turn.events) {
-      if (e.type !== "span.model_request_end") continue;
-      const usage =
-        (e as { model_usage?: { input_tokens?: number; output_tokens?: number; cache_read_input_tokens?: number } }).model_usage ??
-        (e.data as { model_usage?: { input_tokens?: number; output_tokens?: number; cache_read_input_tokens?: number } } | undefined)?.model_usage;
-      if (usage) {
-        input += usage.input_tokens ?? 0;
-        output += usage.output_tokens ?? 0;
-        cacheRead += usage.cache_read_input_tokens ?? 0;
-        calls += 1;
-      }
-    }
-    return { input, output, cacheRead, calls };
-  }, [turn.events]);
-
-  const turnDurationMs =
-    turn.endedAt && turn.triggerTs ? turn.endedAt - turn.triggerTs : totalMs;
-
-  const triggerTitleText = (() => {
-    if (!turn.trigger) return null;
-    const c = (turn.trigger as { content?: Array<{ type: string; text?: string }> }).content;
-    if (!Array.isArray(c)) return null;
-    const t = c.find((b) => b.type === "text")?.text;
-    return t ? t.slice(0, 80) : null;
-  })();
-
-  const borderClass =
-    turn.status === "errored" || turn.status === "terminated"
-      ? "border-danger/50"
-      : "border-border";
-
-  return (
-    <div className={`border ${borderClass} rounded-lg bg-bg-surface/30`}>
-      {/* Header */}
-      <div className="px-4 py-2.5 flex items-center gap-3 text-xs">
-        <button
-          onClick={() => setCollapsed((c) => !c)}
-          className="text-fg-subtle hover:text-fg-muted font-mono w-4 text-center"
-          title={collapsed ? "Expand" : "Collapse"}
-        >
-          {collapsed ? "▸" : "▾"}
-        </button>
-        <span className={`w-1.5 h-1.5 rounded-full shrink-0 ${TRIGGER_DOT[turn.triggerKind]}`} />
-        <span className="font-mono text-fg-muted">{TRIGGER_LABEL[turn.triggerKind]}</span>
-        {triggerTitleText && (
-          <span className="text-fg-subtle truncate max-w-md italic">"{triggerTitleText}"</span>
-        )}
-        <span className="ml-auto flex items-center gap-3 font-mono text-fg-subtle">
-          <span>{spans.length} spans</span>
-          <span>{formatDuration(turnDurationMs)}</span>
-          {tokens.calls > 0 && (
-            <span title={`${tokens.calls} model call${tokens.calls === 1 ? "" : "s"}`}>
-              {tokens.input}↓ {tokens.output}↑
-              {tokens.cacheRead > 0 && ` ⚡${tokens.cacheRead}`}
-            </span>
-          )}
-          <span className={STATUS_TEXT[turn.status]}>{turn.status}</span>
-        </span>
-      </div>
-
-      {!collapsed && spans.length > 0 && (
-        <>
-          {/* Zoom controls — per-card, since each card has its own pxPerMs */}
-          <div className="px-4 pb-2 flex items-center gap-1 text-xs">
-            <button
-              onClick={() => zoomBy(0.5)}
-              className="px-2 py-0.5 rounded border border-border hover:bg-bg-surface text-fg-muted"
-              title="Zoom out"
-            >
-              −
-            </button>
-            <button
-              onClick={resetAuto}
-              className={`px-2 py-0.5 rounded border hover:bg-bg-surface ${mode === "auto" ? "border-info text-info" : "border-border text-fg-muted"}`}
-              title="Auto-pick scale by event density"
-            >
-              auto
-            </button>
-            <button
-              onClick={fitToViewport}
-              className="px-2 py-0.5 rounded border border-border hover:bg-bg-surface text-fg-muted"
-              title="Fit turn duration to viewport"
-            >
-              fit
-            </button>
-            <button
-              onClick={() => zoomBy(2)}
-              className="px-2 py-0.5 rounded border border-border hover:bg-bg-surface text-fg-muted"
-              title="Zoom in"
-            >
-              +
-            </button>
-            <span className="ml-2 font-mono text-fg-subtle">{fmtRate(effectivePxPerMs)}</span>
-          </div>
-
-          {/* Chart — same waterfall layout as before, scoped to this turn */}
-          <div ref={scrollRef} className="overflow-x-auto pb-3 border-t border-border/40">
-            {/* Time axis */}
-            <div className="pt-2 sticky top-0 bg-bg-surface/30 z-10" style={{ width: 224 + chartPx + 80 }}>
-              <div className="flex items-center">
-                <div className="w-56 shrink-0 sticky left-0 bg-bg-surface/30 z-30" />
-                <div className="relative h-5 border-b border-border" style={{ width: chartPx }}>
-                  {ticks.map((t) => (
-                    <div
-                      key={t}
-                      className="absolute top-0 h-full flex flex-col items-start text-[10px] text-fg-subtle font-mono"
-                      style={{ left: `${t * effectivePxPerMs}px` }}
-                    >
-                      <span className="-translate-x-1/2 px-1">{formatDuration(t)}</span>
-                      <div className="w-px flex-1 bg-border" />
-                    </div>
-                  ))}
-                </div>
-                <div className="w-20 shrink-0 sticky right-0 bg-bg-surface/30 z-30" />
-              </div>
-            </div>
-
-            {/* Rows */}
-            {spans.map((s) => {
-              const left = s.startMs * effectivePxPerMs;
-              const width = s.durationMs > 0 ? Math.max(2, s.durationMs * effectivePxPerMs) : 0;
-              const isSelected = selection?.spanKey === s.key;
-              return (
-                <div key={s.key} style={{ width: 224 + chartPx + 80 }}>
-                  <div
-                    className={`flex items-center py-1 border-b border-border/30 hover:bg-bg/40 group cursor-pointer ${isSelected ? "bg-info-subtle/40" : ""}`}
-                    title={
-                      s.detail
-                        ? `${s.label} — ${formatDuration(s.durationMs)} — ${s.detail}`
-                        : `${s.label} — ${formatDuration(s.durationMs)}`
-                    }
-                    onClick={() => onSelectSpan(s)}
-                  >
-                    <div className={`w-56 shrink-0 sticky left-0 group-hover:bg-bg/40 z-20 flex items-center gap-2 text-xs px-4 ${isSelected ? "bg-info-subtle/40" : "bg-bg-surface/30"}`}>
-                      <span className={`w-1.5 h-1.5 rounded-full shrink-0 ${FAMILY_DOT[s.family]}`} />
-                      <div className="min-w-0 flex-1">
-                        <div className="truncate text-fg-muted font-mono">{s.label}</div>
-                        {s.detail && (
-                          <div className="truncate text-fg-subtle font-mono text-[10px]">{s.detail}</div>
-                        )}
-                      </div>
-                    </div>
-                    <div className="relative h-5 shrink-0" style={{ width: chartPx }}>
-                      {width > 0 ? (
-                        <>
-                          <div
-                            className={`absolute h-3 top-1 rounded-sm ${FAMILY_BAR[s.family]} group-hover:opacity-100 opacity-90`}
-                            style={{ left: `${left}px`, width: `${width}px` }}
-                          />
-                          {typeof s.ttftMs === "number" && s.durationMs > 0 && (
-                            <div
-                              className="absolute h-3 top-1 w-px bg-bg-surface"
-                              style={{ left: `${left + s.ttftMs * effectivePxPerMs}px` }}
-                              title={`TTFT ${formatDuration(s.ttftMs)}`}
-                            />
-                          )}
-                        </>
-                      ) : (
-                        <div
-                          className={`absolute top-0 bottom-0 w-px ${FAMILY_DOT[s.family]}`}
-                          style={{ left: `${left}px` }}
-                        />
-                      )}
-                    </div>
-                    <div className={`w-20 shrink-0 sticky right-0 group-hover:bg-bg/40 z-20 text-right text-xs font-mono text-fg-subtle pr-3 ${isSelected ? "bg-info-subtle/40" : "bg-bg-surface/30"}`}>
-                      {s.durationMs > 0 ? formatDuration(s.durationMs) : "·"}
-                    </div>
-                  </div>
-                </div>
-              );
-            })}
-          </div>
-        </>
-      )}
-    </div>
-  );
-}
